@@ -5,9 +5,9 @@ import 'package:flutter_clean_architecture/flutter_clean_architecture.dart';
 
 class HomeController extends Controller {
   int _counter;
-  User _user;
+  User? _user;
   int get counter => _counter;
-  User get user => _user; // data used by the View
+  User? get user => _user; // data used by the View
   final HomePresenter homePresenter;
   // Presenter should always be initialized this way
   HomeController(usersRepo)
@@ -38,7 +38,7 @@ class HomeController extends Controller {
   }
 
   void getUser() => homePresenter.getUser('test-uid');
-  void getUserwithError() => homePresenter.getUser('test-uid231243');
+  void getUserWithError() => homePresenter.getUser('test-uid231243');
 
   void buttonPressed() {
     _counter++;
