@@ -1,10 +1,10 @@
 import './home_controller.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_clean_architecture/flutter_clean_architecture.dart';
+import 'package:flutter_clean_architecture/flutter_clean_architecture.dart' as clean;
 import '../../../data/repositories/data_users_repository.dart';
 
-class HomePage extends View {
-  const HomePage({Key key, this.title}) : super(key: key);
+class HomePage extends clean.View {
+  const HomePage({Key? key, required this.title}) : super(key: key);
 
   final String title;
 
@@ -14,7 +14,7 @@ class HomePage extends View {
       HomePageState();
 }
 
-class HomePageState extends ViewState<HomePage, HomeController> {
+class HomePageState extends clean.ViewState<HomePage, HomeController> {
   HomePageState() : super(HomeController(DataUsersRepository()));
 
   @override
@@ -30,7 +30,7 @@ class HomePageState extends ViewState<HomePage, HomeController> {
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
             children: <Widget>[
-              ControlledWidgetBuilder<HomeController>(
+              clean.ControlledWidgetBuilder<HomeController>(
                 builder: (context, controller) {
                   return Text(
                     'Button pressed ${controller.counter} times.',
@@ -40,15 +40,15 @@ class HomePageState extends ViewState<HomePage, HomeController> {
               const Text(
                 'The current user is',
               ),
-              ControlledWidgetBuilder<HomeController>(
+              clean.ControlledWidgetBuilder<HomeController>(
                 builder: (context, controller) {
                   return Text(
                     controller.user == null ? '' : '${controller.user}',
-                    style: Theme.of(context).textTheme.headline4,
+                    style: Theme.of(context).textTheme.headlineMedium,
                   );
                 },
               ),
-              ControlledWidgetBuilder<HomeController>(
+              clean.ControlledWidgetBuilder<HomeController>(
                 builder: (context, controller) {
                   return ElevatedButton(
                     onPressed: controller.getUser,
@@ -59,10 +59,10 @@ class HomePageState extends ViewState<HomePage, HomeController> {
                   );
                 },
               ),
-              ControlledWidgetBuilder<HomeController>(
+              clean.ControlledWidgetBuilder<HomeController>(
                 builder: (context, controller) {
                   return ElevatedButton(
-                    onPressed: controller.getUserwithError,
+                    onPressed: controller.getUserWithError,
                     child: const Text(
                       'Get User Error',
                       style: TextStyle(color: Colors.white),
@@ -74,7 +74,7 @@ class HomePageState extends ViewState<HomePage, HomeController> {
           ),
         ),
       ),
-      floatingActionButton: ControlledWidgetBuilder<HomeController>(
+      floatingActionButton: clean.ControlledWidgetBuilder<HomeController>(
         builder: (context, controller) {
           return FloatingActionButton(
             onPressed: () => controller.buttonPressed(),
