@@ -1,6 +1,7 @@
-import 'package:example/src/app/pages/home/home_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+
+import 'package:example/src/app/pages/home/home_view.dart';
 
 void main() {
   testWidgets(
