@@ -1,5 +1,6 @@
-import 'package:example/src/data/repositories/data_users_repository.dart';
 import 'package:flutter_test/flutter_test.dart';
+
+import 'package:example/src/data/repositories/data_users_repository.dart';
 
 void main() {
   test('data users repository ...', () async {
